+++ CHANGELOG.md
@@ -1,4 +1,16 @@
 # hapi-rs changelog
+## [0.8.0]
+## Changed
+- AssetLibrary::try_create_first() can now crate nodes other than of Object type.
+- Functions taking optional parent (`Option<NodeHandle>`) are now generic and can take `HoudiniNode` too.
+- Improve the error type handling and printing.
+- Remove `CookOptions` arg from `HoudiniNode::cook`, instead there's new `HoudiniNode::cook_with_options`.
+- Add lots of `debug_assert!` for input validation.
+
+## New
+- `ManagerType` enum represents a network root node.
+- Add several missing geometry APIs.
+
 ## [0.7.0]
 ## Changed
 - Reworked parameter APIs